+++ src/solvers/quasi_Newton.jl
@@ -6,19 +6,22 @@ used with any update rule for the direction.
 
 # Fields
 
-* `p`                 the current iterate, a point on a manifold
-* `X`                 the current gradient
-* `sk`                the current step
-* `yk`                the current gradient difference
-* `direction_update`  an [`AbstractQuasiNewtonDirectionUpdate`](@ref) rule.
-* `retraction_method` an `AbstractRetractionMethod`
-* `stop`              a [`StoppingCriterion`](@ref)
+* `p`                             the current iterate, a point on a manifold
+* `X`                             the current gradient
+* `sk`                            the current step
+* `yk`                            the current gradient difference
+* `direction_update`              an [`AbstractQuasiNewtonDirectionUpdate`](@ref) rule.
+* `nondescent_direction_behavior` a `Symbol` to specify how to handle direction that are not descent ones.
+* `retraction_method`             an `AbstractRetractionMethod`
+* `stepsize`                      a [`Stepsize`](@ref)
+* `stop`                          a [`StoppingCriterion`](@ref)
 
 as well as for internal use
 
-* `p_old`             the last iterate
-* `η`                 the current update direction
-* `X_old`             the last gradient
+* `p_old`                      the last iterate
+* `η`                          the current update direction
+* `X_old`                      the last gradient
+* `nondescent_direction_value` the value from the last inner product check for descent directions
 
 # Constructor
 
@@ -47,6 +50,7 @@ mutable struct QuasiNewtonState{
     S<:Stepsize,
     RTR<:AbstractRetractionMethod,
     VT<:AbstractVectorTransportMethod,
+    R,
 } <: AbstractGradientSolverState
     p::P
     p_old::P
@@ -61,6 +65,7 @@ mutable struct QuasiNewtonState{
     X_old::T
     vector_transport_method::VT
     nondescent_direction_behavior::Symbol
+    nondescent_direction_value::R
 end
 function QuasiNewtonState(
     M::AbstractManifold,
@@ -90,7 +95,7 @@ function QuasiNewtonState(
     VTM<:AbstractVectorTransportMethod,
 }
     sk_init = zero_vector(M, p)
-    return QuasiNewtonState{P,typeof(sk_init),D,SC,typeof(stepsize),RM,VTM}(
+    return QuasiNewtonState{P,typeof(sk_init),D,SC,typeof(stepsize),RM,VTM,Float64}(
         p,
         copy(M, p),
         copy(M, p, initial_vector),
@@ -104,14 +109,27 @@ function QuasiNewtonState(
         copy(M, p, initial_vector),
         vector_transport_method,
         nondescent_direction_behavior,
+        1.0,
     )
 end
 function get_message(qns::QuasiNewtonState)
-    # collect messages from (1) direction update or the (2) the step size and combine them
+    # collect messages from
+    # (1) direction update or the
+    # (2) the step size and combine them
+    # (3) the nondescent behaviour check message
     msg1 = get_message(qns.direction_update)
     msg2 = get_message(qns.stepsize)
-    d = (length(msg1) > 0 && length(msg2) > 0) ? "\n" : "" #divider
-    return "$(msg1)$(d)$(msg2)"
+    msg3 = ""
+    if qns.nondescent_direction_value > 0
+        msg3 = "Computed direction is not a descent direction. The inner product evaluated to $(qns.nondescent_direction_value)."
+        if qns.nondescent_direction_behavior === :step_towards_negative_gradient
+            msg3 = "$(msg3) Resetting to negative gradient."
+        end
+    end
+    d = "$(msg1)"
+    d = "$(length(d) > 0 ? "\n" : "")$(msg2)"
+    d = "$(length(d) > 0 ? "\n" : "")$(msg3)"
+    return d
 end
 function show(io::IO, qns::QuasiNewtonState)
     i = get_count(qns, :Iterations)
@@ -203,7 +221,12 @@ The ``k``th iteration consists of
 * `stopping_criterion`:      ([`StopAfterIteration`](@ref)`(max(1000, memory_size)) | `[`StopWhenGradientNormLess`](@ref)`(1e-6)`)
   specify a [`StoppingCriterion`](@ref)
 * `vector_transport_method`: (`default_vector_transport_method(M, typeof(p))`) a vector transport to use.
-* `nondescent_direction_behavior`: (`:step_towards_negative_gradient`) specify how non-descent direction is handled. Can be either `:ignore` (in which case it is not checked if the selected direction is a descent one) or `:step_towards_negative_gradient` (in which case the direction is replaced with negative gradient).
+* `nondescent_direction_behavior`: (`:step_towards_negative_gradient`) specify how non-descent direction is handled.
+  This can be
+  * ``:step_towards_negative_gradient` – the direction is replaced with negative gradient, a message is stored.
+  * `:ignore` – the check is not performed, so any computed direction is accepted. No message is stored.
+  * any other value performs the check, keeps the direction but stores a message.
+  A stored message can be displayed using [`DebugMessages`](@ref).
 
 # Output
 
@@ -353,12 +376,14 @@ function step_solver!(mp::AbstractManoptProblem, qns::QuasiNewtonState, iter)
     M = get_manifold(mp)
     get_gradient!(mp, qns.X, qns.p)
     qns.direction_update(qns.η, mp, qns)
-    if qns.nondescent_direction_behavior === :step_towards_negative_gradient &&
-        real(inner(M, qns.p, qns.η, qns.X)) > 0
-        # reset direction if not a descent one
-        @warn "Computed direction is not a descent direction; resetting to negative gradient"
-        copyto!(M, qns.η, qns.X)
-        qns.η .*= -1
+    if !(qns.nondescent_direction_behavior === :ignore)
+        qns.nondescent_direction_value = real(inner(M, qns.p, qns.η, qns.X))
+        if qns.nondescent_direction_value > 0
+            if qns.nondescent_direction_behavior === :step_towards_negative_gradient
+                copyto!(M, qns.η, qns.X)
+                qns.η .*= -1
+            end
+        end
     end
     α = qns.stepsize(mp, qns, iter, qns.η)
     copyto!(M, qns.p_old, get_iterate(qns))